requirements.txt
fastapi
uvicorn
redis
slowapi
pydantic-settings
python-dotenv
python-jose
passlib
python-multipart